feat: continue plugin versioning refactoring
