+++ src/AppRouter.tsx
@@ -5,16 +5,16 @@ import Faucets from '@/pages/Faucets';
 import Homepage from '@/pages/Homepage';
 import MySpaces from '@/pages/MySpaces';
 import ShortcutRedirect from '@/pages/ShortcutRedirect';
-import Space from '@/pages/Space';
 import SpaceLauncher from '@/pages/SpaceLauncher';
 import UploadContracts from '@/pages/UploadContracts';
 import Flipper from '@/pages/plugins/Flipper';
 import Polls from '@/pages/plugins/Polls';
 import Posts from '@/pages/plugins/Posts';
 import PendingPosts from '@/pages/plugins/Posts/0.2.x/PendingPosts';
-import Members from '@/pages/space/Members';
-import PendingMembers from '@/pages/space/PendingMembers';
-import Settings from '@/pages/space/Settings';
+import Space from '@/pages/space';
+import Members from '@/pages/space/0.1.x/Members';
+import PendingMembers from '@/pages/space/0.1.x/PendingMembers';
+import Settings from '@/pages/space/0.1.x/Settings';
 import Explorer from 'src/pages/Explorer';
 
 export default function AppRouter() {
