+++ src/pages/Explorer/SpaceCard.tsx
@@ -2,9 +2,9 @@ import { Box, Button, Flex, Heading, Text } from '@chakra-ui/react';
 import { useNavigate } from 'react-router-dom';
 import SpaceAvatar from '@/components/space/SpaceAvatar';
 import useSpace from '@/hooks/useSpace';
-import CancelRequestButton from '@/pages/space/actions/CancelRequestButton';
-import JoinButton from '@/pages/space/actions/JoinButton';
-import SpaceProvider from '@/providers/SpaceProvider';
+import SpaceProvider from '@/pages/space/0.1.x/SpaceProvider';
+import CancelRequestButton from '@/pages/space/0.1.x/actions/CancelRequestButton';
+import JoinButton from '@/pages/space/0.1.x/actions/JoinButton';
 import { OnChainSpace, Props, RegistrationType } from '@/types';
 import pluralize from 'pluralize';
 import { ChainContract } from 'useink';
