+++ src/pages/plugins/Posts/0.2.x/PendingPosts/PendingPostsButton.tsx
@@ -4,8 +4,8 @@ import { MdPendingActions } from 'react-icons/md';
 import { Link as LinkRouter, useLocation } from 'react-router-dom';
 import usePostsContract from '@/hooks/contracts/plugins/usePostsContract';
 import useContractState from '@/hooks/useContractState';
-import { SpacePath } from '@/pages/Space';
-import { useSpaceContext } from '@/providers/SpaceProvider';
+import { SpacePath } from '@/pages/space/0.1.x';
+import { useSpaceContext } from '@/pages/space/0.1.x/SpaceProvider';
 import { useWalletContext } from '@/providers/WalletProvider';
 import { PluginInfo, PostPerm, Props } from '@/types';
 import { stringToNum } from '@/utils/number';
