+++ src/pages/plugins/Posts/index.tsx
@@ -1,19 +1,20 @@
 import { Box, Text } from '@chakra-ui/react';
 import React, { Suspense } from 'react';
-import { useSpaceContext } from '@/providers/SpaceProvider';
+import { useSpacePlugin } from '@/pages/space/0.1.x/SpaceProvider';
 import { PLUGIN_POSTS } from '@/utils/plugins';
+import { compare } from 'compare-versions';
 
 const Posts_V0_1_X = React.lazy(() => import(`./0.1.x/index`));
 const Posts_V0_2_X = React.lazy(() => import(`./0.2.x/index`));
 
 export default function Posts() {
-  const { plugins } = useSpaceContext();
-  const postPlugin = plugins?.find((p) => p.id === PLUGIN_POSTS);
-  if (!postPlugin) {
+  const plugin = useSpacePlugin(PLUGIN_POSTS);
+
+  if (!plugin) {
     return null;
   }
 
-  if (postPlugin.disabled) {
+  if (plugin.disabled) {
     return (
       <Box>
         <Text>This feature is disabled</Text>
@@ -21,12 +22,12 @@ export default function Posts() {
     );
   }
 
-  const { version = '0.1.0' } = postPlugin;
+  const { version = '0.1.0' } = plugin;
 
   return (
     <Suspense fallback={<div>Loading...</div>}>
-      {version.startsWith('0.1.') && <Posts_V0_1_X plugin={postPlugin} />}
-      {version.startsWith('0.2.') && <Posts_V0_2_X plugin={postPlugin} />}
+      {compare(version, '0.2.0', '<') && <Posts_V0_1_X plugin={plugin} />}
+      {compare(version, '0.2.0', '>=') && <Posts_V0_2_X plugin={plugin} />}
     </Suspense>
   );
 }
