+++ src/pages/space/0.1.x/Settings/Membership.tsx
@@ -20,7 +20,7 @@ import { formatBalance } from '@polkadot/util';
 import useCurrentFreeBalance from '@/hooks/space/useCurrentFreeBalance';
 import { useTx } from '@/hooks/useink/useTx';
 import { step2Schema } from '@/pages/SpaceLauncher';
-import { useSpaceContext } from '@/providers/SpaceProvider';
+import { useSpaceContext } from '@/pages/space/0.1.x/SpaceProvider';
 import { Pricing, RegistrationType } from '@/types';
 import { messages } from '@/utils/messages';
 import { notifyTxStatus } from '@/utils/notifications';
