+++ src/pages/space/0.1.x/Settings/Plugins/PostConfig.tsx
@@ -5,7 +5,7 @@ import usePluginContract from '@/hooks/contracts/plugins/usePluginContract';
 import useCurrentFreeBalance from '@/hooks/space/useCurrentFreeBalance';
 import useContractState from '@/hooks/useContractState';
 import { useTx } from '@/hooks/useink/useTx';
-import { useSpaceContext } from '@/providers/SpaceProvider';
+import { useSpaceContext } from '@/pages/space/0.1.x/SpaceProvider';
 import { PluginInfo, PostPerm, Props } from '@/types';
 import { messages } from '@/utils/messages';
 import { notifyTxStatus } from '@/utils/notifications';
