+++ src/pages/space/0.1.x/Settings/Plugins/index.tsx
@@ -3,10 +3,10 @@ import { toast } from 'react-toastify';
 import useCurrentFreeBalance from '@/hooks/space/useCurrentFreeBalance';
 import useContractState from '@/hooks/useContractState';
 import { useTx } from '@/hooks/useink/useTx';
-import PostConfig from '@/pages/space/Settings/Plugins/PostConfig';
-import InstallPluginsButton from '@/pages/space/actions/InstallPluginsButton';
-import UpgradePluginButton from '@/pages/space/actions/UpgradePluginButton';
-import { useSpaceContext } from '@/providers/SpaceProvider';
+import PostConfig from '@/pages/space/0.1.x/Settings/Plugins/PostConfig';
+import { useSpaceContext } from '@/pages/space/0.1.x/SpaceProvider';
+import InstallPluginsButton from '@/pages/space/0.1.x/actions/InstallPluginsButton';
+import UpgradePluginButton from '@/pages/space/0.1.x/actions/UpgradePluginButton';
 import { PluginInfo } from '@/types';
 import { messages } from '@/utils/messages';
 import { notifyTxStatus } from '@/utils/notifications';
