+++ src/pages/space/0.1.x/Settings/UpgradeVersion.tsx
@@ -3,7 +3,7 @@ import { toast } from 'react-toastify';
 import useCurrentFreeBalance from '@/hooks/space/useCurrentFreeBalance';
 import useContractState from '@/hooks/useContractState';
 import { useTx } from '@/hooks/useink/useTx';
-import { useSpaceContext } from '@/providers/SpaceProvider';
+import { useSpaceContext } from '@/pages/space/0.1.x/SpaceProvider';
 import { messages } from '@/utils/messages';
 import { notifyTxStatus } from '@/utils/notifications';
 import { shortenAddress } from '@/utils/string';
