+++ src/pages/space/0.1.x/actions/InstallPluginsButton.tsx
@@ -21,7 +21,7 @@ import { toast } from 'react-toastify';
 import useCurrentFreeBalance from '@/hooks/space/useCurrentFreeBalance';
 import useContractState from '@/hooks/useContractState';
 import { useTx } from '@/hooks/useink/useTx';
-import { useSpaceContext } from '@/providers/SpaceProvider';
+import { useSpaceContext } from '@/pages/space/0.1.x/SpaceProvider';
 import { Plugin } from '@/types';
 import { messages } from '@/utils/messages';
 import { notifyTxStatus } from '@/utils/notifications';
